% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{addLogReturns}
\alias{addLogReturns}
\title{addLogReturns}
\usage{
addLogReturns(df, value)
}
\arguments{
\item{df}{- a data.frame with financial data @seealso getDailyOHLC}

\item{value}{- a single character value (e.g. "Close") or a character vector (e.g. c("Close", "Volume")) to define the value(s) on which to compute the return}
}
\value{
a data.frame with additional columns of the log returns in percent of the specified value(s)
}
\description{
This method computes the log-difference returns in % ( 100 x diff(log(x)) ) and adds it to the input data.frame as a seperate column.
NOTE: First observation cannot have a return value. For computational reasons it has value 0.
}
\examples{

# Get the financial data
prices <- getDailyOHLC("09.2014-11.2014", "DE0007037145")
# Calculate the log returns in percent
log.returns <- addLogReturns(prices, c("Close", "Volume"))

}
