% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{computePerformanceDF}
\alias{computePerformanceDF}
\title{computePerformanceDF}
\usage{
computePerformanceDF(input_df)
}
\arguments{
\item{data.frame}{- with isin and OHLC data in a time period(s)}
}
\value{
data.frame : performanceTable,
}
\description{
This method reads in a data frame with stocks(isin) and their OHLC data and then computes the respective daily performance. E.g. to be plotted in a chart.
}
