% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{formatOHLCTable}
\alias{formatOHLCTable}
\title{Internal private function formats the price data read from the html table into processable data types.
Also format Date into a date class variable
Returns a fuly formatted data.frame row from the values of the html table}
\usage{
formatOHLCTable(inputOHLCTable)
}
\description{
BUG: Cant format turnover
}
