% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{getDailyOHLC}
\alias{getDailyOHLC}
\title{getDailyOHLC}
\usage{
getDailyOHLC(searchDates, isin, input_df)
}
\arguments{
\item{searchDates}{- vector of charachters with different (formats see examples) Date requests: e.g. 20.09.2014-22.09.2014 (normal time period between two dates) // 08.2014 (short for one month) // 2014 (short for one year) // 20.08.2014 (short for one day) // 2013-2014 (time period between two years) // 08.2014-09.2014 (time period between two months)}

\item{isin}{- vector of characters with the id number of the stocks (ISIN = International Stock Identification Number)}

\item{input_df}{- data.frame with dates(as character in format dd.mm.yyyy, e.g. 21.02.2008 (%d.%m.%Y)) and isin(as character, e.g. DE0007037145(RWE AG)). Column names and order --> dates and isin are restricted!}
}
\value{
data.frame : date, isin, open, high, low, close, currency, volume, turnover
}
\description{
This method crawls through the ariva.de website and grabs the daily price data out of the html tables.
You can get OHLC data from specific dates for specific stocks (ISIN required).
Either on putting the different stocks/isins and different dates for prices in a data.frame (input_df). This is good to get different prices on different dates and different stocks.
or by delclaring a vector of dates/intervals and stocks/isins. This is good for comparing the prices on different time periods within a stock or with other stocks.
The date must be in the right format (dd.mm.yyyy) as a character. It is not checked if the date is a correct trading day.
It is also not checked if the date intervals are correct.

Status Quo:
- Only Xetra price data
- price data is cleaned up from stock splits (clean_splits)

The method strictly depends on the website structure of ariva and GET request of the csv file!
}
\examples{

# First example for the first search option (input parameter: only input_df)
#' # Build the input parameter data.frame
dates <- c("25.02.2008","27.02.2008")
isin <- c("DE0007037145","DE0007037145")
input_df <- data.frame(dates, isin)

# Get the prices for RWE AG (DE0007037145) of the 25.02.2008 and 27.02.2008
prices1 <- getDailyOHLC(input_df = input_df)

# Second example for the second search option (input parameter: searchDates and isin)
# Build the input parameters
isin  <- c("DE0007037145","DE0007664039")
searchDates  <- c("20.09.2014-22.09.2014","20.08.2014","06.05.2014-16.05.2014")

# Get the prices for RWE AG (DE0007037145) and VW VZ (DE0007664039) for different time periods or dates (see searchDates parameter above)
prices2 <- getDailyOHLC(searchDates=searchDates, isin=isin)


}
