% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{getISINFromYahooTicker}
\alias{getISINFromYahooTicker}
\title{getISINFromYahooTicker}
\usage{
getISINFromYahooTicker(ytickers)
}
\arguments{
\item{vector}{of character - Yahoo Ticker (e.g. ADS.DE)}
}
\value{
data.frame : yahooTicker , isin OF THE UNIQUE INPUT VECTOR !!
}
\description{
This function takes an vector of Yahoo Ticker characters and returns the added ISIN number.
}
