% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{getOHLC}
\alias{getOHLC}
\title{getOHLC}
\usage{
getOHLC(searchDates, isin, input_df)
}
\arguments{
\item{input_df}{- data.frame with dates(as character in format dd.mm.yyyy e.g. 21.02.2008 (%d.%m.%Y)) and isin(as character - international stock identification number (e.g. DE0007037145 // RWE AG)
searchDate = vector of charachters with Date request e.g. 20.09.2014-22.09.2014 // 08.2014 // 2014 // 20.08.2014 // 2013-2014 // 08.2014-09.2014
isin = vector of the id number of the stocks}
}
\value{
data.frame : date,isin,open,high,low,close,currency,volume,turnover
}
\description{
This method crawls through the ariva.de financal website and grabs price data out of the html tables.
You can get OHLC data from specific dates for specific stocks (ISIN required).
Either on putting the different stocks/isins and different dates for prices in a data.frame (input_df). This is good to get different prices on different dates and different stocks.
or by delclaring a vector of dates/intervals and stocks/isins. This is good for comparing the prices on different time periods within a stock or with other stocks.
The date must be in the right format (dd.mm.yyyy) as a character. It is not checked if the date is a correct trading day.
It is also not checked if the date intervals are correct.

The method strictly depends on the website structure of ariva!

Needs libraries RCurl, XML and stringr

Ex. 1: dates <- c("25.02.2008","27.02.2008"); isin <- c("DE0007037145","DE0007037145"); input_df <- data.frame(dates,isin); getOHLC(input_df=input_df)
Ex. 2: isin  <- c("DE0007037145","DE0007664039") searchDates  <- c("20.09.2014-22.09.2014","20.08.2014","06.05.2014-16.05.2014"); getOHLC(searchDates=searchDates, isin=isin)
}
