% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{getStockInfoFromIsin}
\alias{getStockInfoFromIsin}
\title{getStockInfoFromIsin}
\usage{
getStockInfoFromIsin(isin)
}
\arguments{
\item{isin}{- vector of charachters with different stocks (isin), e.g. DE0007037145(RWE AG)}
}
\value{
data.frame with stock infos
}
\description{
This method crawls through the ariva.de website and scrapes out information about the given stocks (isin).
Status Quo: ticker info, sector.

NOTE: Only the column/variable NominalValue is checked for NA values (== "-")

The method strictly depends on the website structure of ariva!
}
\examples{

# Build the input parameter
isin <- c("DE0007037145", "DE000ENAG999")

# Get the information about for RWE AG (DE0007037145)
stockInfos <- getStockInfoFromIsin(isin = isin)

}
