% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{getStockMarketIndexInfoFromLink}
\alias{getStockMarketIndexInfoFromLink}
\title{getStockMarketIndexInfoFromLink}
\usage{
getStockMarketIndexInfoFromLink(link)
}
\arguments{
\item{link}{- link of the ariva website to adress the main page of the stock market index}
}
\value{
data.frame with stock market index info
}
\description{
This method crawls through the ariva.de website and scrapes out the stock market index info given its link @seealso getStockMarketIndexList

The method strictly depends on the website structure of ariva!
}
\examples{

link <-  "http://www.ariva.de/dax-30"
# Get a list of stock market indexes from ariva.de
daxInfo <- getStockMarketIndexInfoFromLink(link)

}
