% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{getStockMarketIndexList}
\alias{getStockMarketIndexList}
\title{getStockMarketIndexList}
\usage{
getStockMarketIndexList(isin)
}
\value{
data.frame with stock market indexes
}
\description{
This method crawls through the ariva.de website and scrapes out the stock market indexes

The method strictly depends on the website structure of ariva!
}
\examples{

# Get a list of stock market indexes from ariva.de
indexList <- getStockMarketIndexList()

}
