% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{groupInputData}
\alias{groupInputData}
\title{DEPRICATED}
\usage{
groupInputData(dataFrame)
}
\description{
private function for @seealso getDailyOHLC. Prepares the input data.frame of dates and corresponding isin.
Groups the data to reduce the request on ariva.de
Returns a list with the grouped data.frames.
Grouping: by year then month and then stock/isin
}
