% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{readHTMLPriceTable}
\alias{readHTMLPriceTable}
\title{This private method helps the getOHLC function to read out the price table on ariva website.
Therefore it gets a data.frame of a stock with the dates for the required prices and returns the prices data.frame.}
\usage{
readHTMLPriceTable(df)
}
\description{
This private method helps the getOHLC function to read out the price table on ariva website.
Therefore it gets a data.frame of a stock with the dates for the required prices and returns the prices data.frame.
}
