% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{setLogging}
\alias{setLogging}
\title{setLogging}
\usage{
setLogging(logger)
}
\arguments{
\item{logger}{- A boolean variable. TRUE for printing out the logging outputs in the console.}
}
\description{
This function sets a global options variable called "logging" to TRUE OR FALSE. By default it is FALSE, so no logging is displayed.
}
